feat: adiciona novos testes

+++ src/main/java/dev/franke/felipe/compras/compras/api/service/CompradorService.java
@@ -1,11 +1,7 @@
 package dev.franke.felipe.compras.compras.api.service;
 
 import dev.franke.felipe.compras.compras.api.dto.in.CompradorINDTO;
-import dev.franke.felipe.compras.compras.api.exception.CompradorINObrigatorioException;
-import dev.franke.felipe.compras.compras.api.exception.CompradorNaoEncontradoException;
-import dev.franke.felipe.compras.compras.api.exception.CompradorObrigatorioException;
-import dev.franke.felipe.compras.compras.api.exception.IdCompradorInvalidoException;
-import dev.franke.felipe.compras.compras.api.exception.ValorCompraInvalidoException;
+import dev.franke.felipe.compras.compras.api.exception.*;
 import dev.franke.felipe.compras.compras.api.mapper.CompradorMapper;
 import dev.franke.felipe.compras.compras.api.model.Comprador;
 import dev.franke.felipe.compras.compras.api.repository.CompradorRepository;
@@ -35,6 +31,11 @@ public class CompradorService {
         return this.compradorRepository.compradoresNegativados();
     }
 
+    private static void validaCompradorERequisicao(Comprador comprador, BigDecimal valor) {
+        if (comprador == null) throw new CompradorObrigatorioException("O comprador e obrigatorio");
+        if (valor == null) throw new ValorProdutoObrigatorioException("O valor e obrigatorio");
+    }
+
     public Comprador porId(String id) {
         try {
             return this.compradorRepository
@@ -47,6 +48,7 @@ public class CompradorService {
     }
 
     public Comprador cadastra(CompradorINDTO requisicao) {
+        if (requisicao == null) throw new CompradorINObrigatorioException("A requisicao e obrigatoria");
         requisicao.validaTudo();
         var comprador = MAPPER.compradorINDTOParaComprador(requisicao);
         return compradorRepository.save(comprador);
@@ -61,17 +63,20 @@ public class CompradorService {
     }
 
     public Comprador atualizaSaldoDebito(Comprador comprador, BigDecimal valor) {
+        validaCompradorERequisicao(comprador, valor);
         comprador.setSaldoDebito(valor);
         return compradorRepository.save(comprador);
     }
 
     public Comprador atualizaSaldoValeAlimentacao(Comprador comprador, BigDecimal valor) {
+        validaCompradorERequisicao(comprador, valor);
         comprador.setSaldoValeAlimentacao(valor);
         return compradorRepository.save(comprador);
     }
 
     public Comprador atualizaTotalCompras(Comprador comprador, int valor) {
-        if (valor <= 0 || valor < comprador.getTotalCompras().intValue()) {
+        if (comprador == null) throw new CompradorObrigatorioException("O comprador e obrigatorio");
+        if (valor <= 0 || valor < comprador.getTotalCompras()) {
             throw new ValorCompraInvalidoException("Valor invalido: " + valor);
         }
         comprador.setTotalCompras(valor);
@@ -79,6 +84,7 @@ public class CompradorService {
     }
 
     public void apaga(Comprador comprador) {
+        if (comprador == null) throw new CompradorObrigatorioException("O comprador e obrigatorio");
         compradorRepository.delete(comprador);
     }
 }
